[refactor] UserEntity의 패키지를 변경 #20
