+++ src/main/kotlin/dev/codeswamp/core/article/application/service/ArticleApplicationService.kt
@@ -2,7 +2,7 @@ package dev.codeswamp.core.article.application.service
 
 import dev.codeswamp.core.article.presentation.dto.request.ArticleMetadataDto
 import dev.codeswamp.core.article.presentation.dto.response.ArticleReadResponseDto
-import dev.codeswamp.core.user.infrastructure.entity.UserEntity
+import dev.codeswamp.core.user.infrastructure.persistence.entity.UserEntity
 
 interface ArticleApplicationService {
