+++ src/main/kotlin/dev/codeswamp/core/article/application/service/ArticleApplicationServiceImpl.kt
@@ -4,7 +4,7 @@ import dev.codeswamp.core.article.domain.service.ArticleService
 import dev.codeswamp.core.article.infrastructure.search.ArticleSearchRepository
 import dev.codeswamp.core.article.presentation.dto.request.ArticleMetadataDto
 import dev.codeswamp.core.article.presentation.dto.response.ArticleReadResponseDto
-import dev.codeswamp.core.user.infrastructure.entity.UserEntity
+import dev.codeswamp.core.user.infrastructure.persistence.entity.UserEntity
 import org.springframework.stereotype.Service
 
 @Service
