+++ src/main/kotlin/dev/codeswamp/core/article/infrastructure/persistence/entity/ArticleMetadataEntity.kt
@@ -2,7 +2,7 @@ package dev.codeswamp.core.article.infrastructure.persistence.entity
 
 import dev.codeswamp.core.comment.entity.Comment
 import dev.codeswamp.core.folder.domain.entity.Folder
-import dev.codeswamp.core.user.infrastructure.entity.UserEntity
+import dev.codeswamp.core.user.infrastructure.persistence.entity.UserEntity
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
